+++ lib/util/themes.dart
@@ -19,6 +19,13 @@ class Themes {
     ),
   );
 
+  Widget charitiesBarTitle = new Text(
+    'Charities',
+    style: new TextStyle(
+      color: Colors.limeAccent[700],
+    ),
+  );
+
   Icon searchIcon = new Icon(Icons.search, color: Colors.limeAccent[700],);
 
   Icon backIcon = new Icon(Icons.arrow_back, color: Colors.limeAccent[700],);